+++ packages/storage/src/AwsClients/S3/runtime/index.native.ts
@@ -1,0 +1,14 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+// Entry point for ReactNative-specific S3 client utilities
+export {
+	SEND_DOWNLOAD_PROGRESS_EVENT,
+	SEND_UPLOAD_PROGRESS_EVENT,
+	CANCELED_ERROR_MESSAGE,
+	CONTENT_SHA256_HEADER,
+} from './constants';
+export { s3TransferHandler } from './s3TransferHandler/xhr';
+export { parser } from './xmlParser/pureJs';
+export { isCancelError } from './xhrTransferHandler';
+export { toBase64, utf8Encode } from './base64/index.native';