+++ packages/storage/src/common/StorageConstants.ts
@@ -1,10 +1,10 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
 export const AMPLIFY_SYMBOL = (
 	typeof Symbol !== 'undefined' && typeof Symbol.for === 'function'
 		? Symbol.for('amplify_default')
 		: '@@amplify_default'
 ) as Symbol;
-
-export const SET_CONTENT_LENGTH_HEADER = 'contentLengthMiddleware';
 
 export const localTestingStorageEndpoint = 'http://localhost:20005';