+++ packages/aws-amplify/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.3.6](https://github.com/aws-amplify/amplify-js/compare/aws-amplify@5.3.5...aws-amplify@5.3.6) (2023-07-31)
+
+**Note:** Version bump only for package aws-amplify
+
+## [5.3.5](https://github.com/aws-amplify/amplify-js/compare/aws-amplify@5.3.4...aws-amplify@5.3.5) (2023-07-20)
+
+**Note:** Version bump only for package aws-amplify
+
+## [5.3.4](https://github.com/aws-amplify/amplify-js/compare/aws-amplify@5.3.3...aws-amplify@5.3.4) (2023-07-13)
+
+**Note:** Version bump only for package aws-amplify
+
+## [5.3.3](https://github.com/aws-amplify/amplify-js/compare/aws-amplify@5.3.2...aws-amplify@5.3.3) (2023-06-28)
+
+**Note:** Version bump only for package aws-amplify
+
+## [5.3.2](https://github.com/aws-amplify/amplify-js/compare/aws-amplify@5.3.1...aws-amplify@5.3.2) (2023-06-27)
+
+**Note:** Version bump only for package aws-amplify
+
 ## [5.3.1](https://github.com/aws-amplify/amplify-js/compare/aws-amplify@5.3.0...aws-amplify@5.3.1) (2023-06-21)
 
 **Note:** Version bump only for package aws-amplify