+++ packages/datastore/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.7.0](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/datastore@4.6.5...@aws-amplify/datastore@4.7.0) (2023-07-31)
+
+### Features
+
+- custom user agent Geo changes for UI handoff ([#11632](https://github.com/aws-amplify/amplify-js/issues/11632)) ([01bfa8f](https://github.com/aws-amplify/amplify-js/commit/01bfa8f692737bd14422f7dc2eae11ed00c19048))
+
+## [4.6.5](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/datastore@4.6.4...@aws-amplify/datastore@4.6.5) (2023-07-20)
+
+**Note:** Version bump only for package @aws-amplify/datastore
+
+## [4.6.4](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/datastore@4.6.3...@aws-amplify/datastore@4.6.4) (2023-07-13)
+
+**Note:** Version bump only for package @aws-amplify/datastore
+
+## [4.6.3](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/datastore@4.6.2...@aws-amplify/datastore@4.6.3) (2023-06-28)
+
+**Note:** Version bump only for package @aws-amplify/datastore
+
+## [4.6.2](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/datastore@4.6.1...@aws-amplify/datastore@4.6.2) (2023-06-27)
+
+**Note:** Version bump only for package @aws-amplify/datastore
+
 ## [4.6.1](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/datastore@4.6.0...@aws-amplify/datastore@4.6.1) (2023-06-21)
 
 ### Bug Fixes